Separate Presenter into 2 for AddToHomeScreen and Open Link With
